Deal better with null value at getBigDecimal in BitVectorAccessor

+++ java/flight/flight-jdbc-driver/src/main/java/org/apache/arrow/driver/jdbc/accessor/impl/numeric/ArrowFlightJdbcBitVectorAccessor.java
@@ -104,8 +104,9 @@ public class ArrowFlightJdbcBitVectorAccessor extends ArrowFlightJdbcAccessor {
 
   @Override
   public BigDecimal getBigDecimal() {
-    final BigDecimal value = BigDecimal.valueOf(this.getLong());
-    return this.wasNull ? null : value;
+    final long value = this.getLong();
+
+    return this.wasNull ? null : BigDecimal.valueOf(value);
   }
 
   @Override
